Add register a new account UI
