+++ frontend/components/Footer.tsx
@@ -3,7 +3,7 @@ import { Container } from "react-bootstrap";
 const Footer = (): JSX.Element => (
   <footer className="mb-3">
     <Container fluid>
-      Tania 1.7.0. Made for the ♥ of plants &copy; 2021 Copyright.
+      Tania is made for the ♥ of plants &copy; 2022.
     </Container>
   </footer>
 );
