+++ frontend/pages/index.tsx
@@ -1,87 +1,110 @@
-import { useState } from "react";
 import type { NextPage } from "next";
-import Image from "next/image";
-import { useRouter } from "next/router";
-import { Card, Col, Container, Form, Row } from "react-bootstrap";
-import { FaUnlock } from "react-icons/fa";
+import Link from "next/link";
+import { Col, Form, Nav, Row, Table } from "react-bootstrap";
 
-import ButtonIcon from "../components/ButtonIcon";
+import Layout from "@/components/Layout";
+import Panel from "@/components/Panel";
+import TableTaskItem from "@/components/TableTaskItem";
+import { dashboardData, cropsData, tasksData } from "@/data/index";
 
-const Home: NextPage = () => {
-  const router = useRouter();
-  const [username, setUsername] = useState("");
-  const [password, setPassword] = useState("");
-  const [usernameIsEmpty, setUsernameIsEmpty] = useState(false);
-  const [passwordIsEmpty, setPasswordIsEmpty] = useState(false);
-
-  const handleSubmit = () => {
-    if (!username || !password) {
-      setUsernameIsEmpty(!username);
-      setPasswordIsEmpty(!password);
-    } else {
-      router.push("dashboard");
-    }
-  };
-
+const Root: NextPage = () => {
   return (
-    <div className="bg-gray d-flex align-items-center vh-100">
-      <Container fluid>
-        <Row>
-          <Col sm={12} md={{ span: 6, offset: 3 }} lg={{ span: 4, offset: 4 }}>
-            <Card>
-              <Card.Body>
-                <div className="text-center">
-                  <Image
-                    src={"/images/logobig.png"}
-                    layout="fixed"
-                    width={200}
-                    height={65}
-                  />
+    <Layout>
+      <Row>
+        <Col>
+          <h3 className="pb-3">Dashboard</h3>
+        </Col>
+      </Row>
+      <Row>
+        <Panel
+          title="At A Glance"
+          md={6}
+          lg={6}
+        >
+          <Nav className="flex-column">
+            {dashboardData.map(({ name, route, icon }) => (
+              <Nav.Item key={name} className="mb-1">
+                <div className="d-flex align-items-center">
+                  {icon}
+                  <Link href={route}>{name}</Link>
                 </div>
-                <Form>
-                  <Form.Group className="mb-3">
-                    <Form.Label>Username</Form.Label>
-                    <Form.Control
-                      type="text"
-                      placeholder="Input your username here"
-                      onChange={(e) => setUsername(e.target.value)}
-                    />
-                    {usernameIsEmpty && (
-                      <Form.Text className="text-danger">
-                        The username field is required
-                      </Form.Text>
-                    )}
-                  </Form.Group>
-                  <Form.Group className="mb-3">
-                    <Form.Label>Password</Form.Label>
-                    <Form.Control
-                      type="password"
-                      placeholder="Your password here"
-                      onChange={(e) => setPassword(e.target.value)}
-                    />
-                    {passwordIsEmpty && (
-                      <Form.Text className="text-danger">
-                        The password field is required
-                      </Form.Text>
-                    )}
-                  </Form.Group>
-                </Form>
-                <div className="text-center">
-                  <ButtonIcon
-                    label="Login"
-                    icon={<FaUnlock className="me-1" />}
-                    type="submit"
-                    variant="primary"
-                    onClick={handleSubmit}
-                  />
-                </div>
-              </Card.Body>
-            </Card>
-          </Col>
-        </Row>
-      </Container>
-    </div>
+              </Nav.Item>
+            ))}
+          </Nav>
+        </Panel>
+        <Panel title="What is on Production" md={6} lg={6}>
+          <>
+            <div className="mb-3">
+              <Link href="/crops">See all Crops</Link>
+            </div>
+            <Table responsive>
+              <thead>
+                <tr>
+                  <th>Varieties</th>
+                  <th>Qty</th>
+                </tr>
+              </thead>
+              <tbody>
+                {cropsData &&
+                  cropsData.map(({ id, varieties, qty }) => (
+                    <tr key={id}>
+                      <td>
+                        <Link href={`/crops/${id}`}>{varieties}</Link>
+                      </td>
+                      <td>{qty}</td>
+                    </tr>
+                  ))}
+              </tbody>
+            </Table>
+          </>
+        </Panel>
+      </Row>
+      <Row>
+        <Panel title="Tasks">
+          <>
+            <div className="mb-3">
+              <Link href="/tasks">See all Tasks</Link>
+            </div>
+            <Table responsive>
+              <thead>
+                <tr>
+                  <th></th>
+                  <th className="w-75">Items</th>
+                  <th>Category</th>
+                </tr>
+              </thead>
+              <tbody>
+                {tasksData &&
+                  tasksData.map(
+                    ({ id, item, details, dueDate, priority, category }) => (
+                      <tr key={id}>
+                        <td>
+                          <Form>
+                            <Form.Check type="checkbox" />
+                          </Form>
+                        </td>
+                        <td>
+                          <TableTaskItem
+                            id={id}
+                            item={item}
+                            details={details}
+                            dueDate={dueDate}
+                            priority={priority}
+                          />
+                        </td>
+                        <td>
+                          <span className="text-uppercase">{category}</span>
+                        </td>
+                      </tr>
+                    )
+                  )}
+              </tbody>
+            </Table>
+          </>
+        </Panel>
+      </Row>
+    </Layout>
   );
 };
 
-export default Home;
+export default Root;
