+++ frontend/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "baseUrl": ".",
     "target": "es5",
     "lib": [
       "dom",
@@ -18,7 +19,12 @@
     "isolatedModules": true,
     "jsx": "preserve",
     "pretty": true,
-    "incremental": true
+    "incremental": true,
+    "paths": {
+      "@/components/*": ["components/*"],
+      "@/data/*": ["data/*"],
+      "@/types/*": ["types/*"]
+    },
   },
   "include": [
     "next-env.d.ts",
